+++ server/migrations/down/005_request_rate_limits.sql
@@ -1,0 +1,6 @@
+BEGIN;
+
+DROP TABLE IF EXISTS request_rate_limits;
+
+DELETE FROM schema_migrations WHERE version = 5;
+COMMIT;