+++ server/migrations/down/008_session_revocations.sql
@@ -1,0 +1,6 @@
+BEGIN;
+
+DROP TABLE IF EXISTS user_session_revocations;
+
+DELETE FROM schema_migrations WHERE version = 8;
+COMMIT;